vim: Fix a bug where WinDo() might lead to enter the insert mode
When a terminal window is open, executing WinDo(...) will result
in vim entering the insert mode on the *wrong* buffer where WinDo
was invoked (possibly a bug of neovim). For example, <leader>ha
(GitGutterStageHunk) is followed by entering the insert mode. This
is a side-effect of an auto-startinsert autocmd for terminal buffers.

A workaround is to use `noautocmd` modifier to prevent such autocmds
from firing.