vim: Avoid errors when loading non-arm64 python3 host on a M1 mac
On a M1 mac, it is possible that neovim runs under rosetta2
(i.e. x86_64 binary) but the python3 detected and being used is
actually an arm64 binary. For example, as of today arm64 neovim cannot
build from source due to some dependency problems. In such cases,
python3 host cannot be loaded and a bunch of annoying errors will occur.

To avoid these errors, we detect whether the python3 host and neovim
binaries have the same architecture. If there is a mismatch, we could
show a warning message and then fall back to /usr/local/bin/python3
(typically a intel-homebrew x86_64 binary as per the new convention).